added userscreen and bugs fixes

+++ app/src/main/java/com/boris/expert/csvmagic/model/Feature.kt
@@ -5,9 +5,9 @@ data class Feature(
     val name: String,
     val credit_price: Int,
     var duration: Int,
-    var memory: Int,
+    var memory: Float,
     var createdAt:Long,
     var expiredAt:Long
 ){
-    constructor():this(0,"",0,0,0,0,0)
+    constructor():this(0,"",0,0,0F,0,0)
 }
